requirements.txt
appdirs==1.4.3
coverage==4.3.4
Django==1.9.9
django-multiselectfield==0.1.5
djangorestframework==3.4.1
olefile==0.44
mysqlclient==1.3.7
dj-database-url==0.4.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
whitenoise==3.2.1
